refactor: 책장 컴포넌트 추상화
